requirements.txt
# Bid Evaluation Agent POC - Requirements
# REQUIRES Python 3.10-3.13 (ChromaDB not compatible with 3.14)

# Core
streamlit
groq
tavily-python
python-dotenv
requests
pandas

# Observability
opentelemetry-api
opentelemetry-sdk

# Vector Database & Embeddings (Historical RAG)
chromadb
sentence-transformers
